Backends: SDL3: Fix macOS mouse position incorrect for fullscreen apps
This commit fixes an issue with macOS on devices with a notch. SDL3
no longer provides a client area that encompasses the notch, so any
reporting of the mouse location must factor this in for fullscreen
SDL3 apps on macOS. SDL3 developers suggested using SDL_GetMouseState()
to retrieve the correct location until they can find a solution for
rendering into the notch area.